+++ princexmlserver/__init__.py
@@ -28,6 +28,7 @@ def main(global_config, **settings):
     config.add_static_view('static', 'static', cache_max_age=3600)
     config.add_route('home', '/')
     config.add_route('convert', '/convert')
+    config.add_route('ready', '/ready')
     config.include('pyramid_chameleon')
     config.scan()
     return config.make_wsgi_app()